pyproject.toml

[project]
name = "testsuite"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12,<3.13"

dependencies = [
    #    "orb-models>=0.5.5",
    "fairchem-core==1.10.0",
    "phonopy>=2.43.2",
    "pymatgen>=2025.6.14",
    "gpaw>=25.7.0",
    "nequip-allegro>=0.7.1",
    "mace-torch",
    "matscipy>=1.1.1",
    "nequix>=0.2.0",
    "sevenn>=0.11.2.post1",
    "seekpath>=2.1.0",
    "phono3py>=3.19.2",
    "np>=1.0.2",
    "scipy-stubs==1.16.2.0",
    "seaborn>=0.13.2",
    "pandas-stubs==2.3.2.250827",
    #    "mattersim>=1.2.0",
    "ptitprince>=0.3.1",
    "scikit-image>=0.25.2",
    "torch-scatter==2.1.2",
    "torchvision==0.19.1",
    "torch_sparse==0.6.18",
    "torch-geometric==2.7.0",
]

[tool.uv.sources]
mace-torch = { git = "https://github.com/Jpx3/mace.git" }
#matscipy = { git = "https://github.com/Jpx3/matscipy.git", branch = "25_numpy_2"}

[tool.uv]
find-links = ["https://data.pyg.org/whl/torch-2.4.1+cu121.html", "https://data.pyg.org/whl/torch-2.4.1+cpu.html"]
